refactoring BrowseProductsPage to use react query and split components inside
