+++ package.json
@@ -11,7 +11,8 @@
     "server": "json-server --watch src/data/db.json --delay 500",
     "start": "concurrently \"npm run server\" \"npm run dev\"",
     "test": "vitest",
-    "test:ui": "vitest -- ui"
+    "test:ui": "vitest -- ui",
+    "coverage": "vitest run --coverage"
   },
   "dependencies": {
     "@auth0/auth0-react": "^2.2.3",
@@ -43,6 +44,7 @@
     "@typescript-eslint/eslint-plugin": "^6.14.0",
     "@typescript-eslint/parser": "^6.14.0",
     "@vitejs/plugin-react": "^4.2.1",
+    "@vitest/coverage-v8": "^3.2.4",
     "autoprefixer": "^10.4.16",
     "concurrently": "^8.2.2",
     "eslint": "^8.55.0",
